adjust icon import

+++ lib/renew_collab/symbol.ex
@@ -24,12 +24,15 @@ defmodule RenewCollab.Symbol do
 
   """
   def list_shapes do
-    Repo.delete_all(Shape)
 
-    RenewCollab.Symbol.PredefinedSymbols.all()
-    |> Enum.each(fn shape ->
-      %Shape{} |> Shape.changeset(shape) |> Repo.insert() |> dbg()
-    end)
+    if Repo.aggregate(Shape, :count) < 5 do
+      Repo.delete_all(Shape) 
+      
+      RenewCollab.Symbol.PredefinedSymbols.all()
+      |> Enum.each(fn shape ->
+        %Shape{} |> Shape.changeset(shape) |> Repo.insert() |> dbg()
+      end)
+    end
 
     Repo.all(Shape)
     |> Repo.preload(
